requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
opencv-python==4.8.1.78
pandas==2.1.3
numpy>=1.24
Pillow==9.5.0
gunicorn==21.2.0
Werkzeug==3.0.1
APScheduler==3.10.4
python-dotenv==1.0.0
SQLAlchemy==2.0.23
